% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDailyMaps.R
\name{calc_risk}
\alias{calc_risk}
\title{Calculate risk}
\usage{
calc_risk(I, g, pop)
}
\arguments{
\item{I}{Number infected}

\item{g}{Event size}

\item{pop}{Population size}
}
\description{
Calculate risk
}
