% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county}
\alias{county}
\title{USA county geoms}
\format{
Simple feature collection with 6 features and 3 fields:
\describe{
  \item{GEOID}{FIPS number}
  \item{NAME}{County common name}
  \item{stname}{Abbreviation of state the county is in}
  \item{geometry}{Polygon geometry for mapping}
}
}
\usage{
county
}
\description{
A simple feature collection of USA county geoms for use in Leaflet
}
\keyword{datasets}
