% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop}
\alias{pop}
\title{USA county population.}
\format{
A data frame with 3221 rows and 2 variables:
\describe{
  \item{fips}{FIPS number}
  \item{pop}{Population}
}
}
\usage{
pop
}
\description{
A dataset containing FIPS and county population.
}
\keyword{datasets}
