% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stateline}
\alias{stateline}
\title{USA state geoms}
\format{
Simple feature collection with 6 features and 2 fields:
\describe{
  \item{stname}{Abbreviation of state}
  \item{name}{State full name}
  \item{geomtry}{Polygon geometry for mappings}
}
}
\usage{
stateline
}
\description{
A simple feature collection of USA state geoms for use in Leaflet.
}
\keyword{datasets}
