% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDailyMaps.R
\name{create_c19r_data}
\alias{create_c19r_data}
\title{Create covid19 risk prediction data}
\usage{
create_c19r_data(
  output = "usa_risk_counties.csv",
  event_size = c(10, 15, 20, 25, 50, 100, 500, 1000, 5000),
  asc_bias_list = c(3, 4, 5),
  scale_factor = (10/14)
)
}
\arguments{
\item{output}{Output file}

\item{event_size}{Vector of event sizes to predict risk with}

\item{asc_bias_list}{Vector of ascertainment biases to predict risk with}

\item{scale_factor}{Scaling factor, should be =< 1}
}
\description{
Create covid19 risk prediction data
}
